requirements.txt
keyboard~=0.13.5
Faker~=11.3.0
kafka-python~=2.0.2
geopy~=2.2.0